requirements.txt
django
gunicorn
whitenoise
djangorestframework
psycopg2
dj-database-url
coreapi
coreschema